Separate bandit and slates tests
